Update README.md
Breast cancer screening is an important strategy to allow for early detection and ensure a greater probability of having a good outcome in treatment. The research goal of this project is to utilize a publicly available dataset to identify and validate a generic blood screening tests for early breast cancer detection in the general population. In this project, we will assess how models based on the input variables collected during routine blood analysis may be used to predict the presence of breast cancer. Furthermore, the applicability of the predictive model to larger and more heterogeneous populations can be subsequently assessed. Since this is a bi-class classification problem, supervised learning algorithms that classify observations based on a set of input variables will be used to assess the suitability of the algorithm for predicting breast cancer in the given dataset.